Refactor CLI operator actions into their own files
